pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dagster-dag-factory"
version = "0.0.1"
dependencies = [
    "dagster==1.11.1",
    "dagster-webserver==1.11.1",
    "croniter>=2.0.0",
    "pyodbc>=5.0.0",
    "snowflake-connector-python>=3.12.0",
    "dagster-snowflake>=0.23.0",
    "boto3>=1.34.0",
    "pandas>=2.2.0",
    "paramiko>=3.4.0",
    "pysftp==0.2.9",
    "pyyaml>=6.0",
    "pydantic>=2.7.0",
    "cryptography>=42.0.0",
    "pendulum>=3.0.0",
    "psycopg2-binary>=2.9.0"
]

[tool.setuptools]
packages = ["dagster_dag_factory"]
package-dir = {"" = "src"}
